+++ librtl/src/str_pos_extr.c
@@ -61,7 +61,7 @@
  *
  */
 int str$pos_extr(struct dsc$descriptor_s* destination_string, const struct dsc$descriptor_s* source_string,
-        const long* start_position, const long* end_position)
+        const int* start_position, const int* end_position)
 {
     char* s2_ptr; /* Pointer to second string */
     unsigned short s2_length; /* Length of second string */
